% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlstrings.R
\name{generate_sql_strings}
\alias{generate_sql_strings}
\title{Read all files from path and return a list object with attributes mapped to sql statements and queries}
\usage{
generate_sql_strings(path = "")
}
\arguments{
\item{path}{\itemize{
\item a path to a folder or a file containing sql code
}}
}
\value{
\itemize{
\item a list with attributes mapped to sql statements
}
}
\description{
Read all files from path and return a list object with attributes mapped to sql statements and queries
}
